New unique identifier, new name attribute on user, updated style of login and signup
