+++ client/src/components/chat/MessageContainer.vue
@@ -24,17 +24,12 @@ export default {
     },
     components: {
       Message
-    },
-    created() {
-      
     }
 }
 </script>
 
 <style>
 .message-container {
-  overflow-y: scroll;
   margin-bottom: 100px;
-  /* flex: 1; */
-  /* max-height: calc(100vh - ); */
 }
+</style>
