+++ server/app/schemas/__init__.py
@@ -1,8 +1,10 @@
 from app.schemas.role import RoleSchema
-from app.schemas.user import LoginSchema, UserSchema
+from app.schemas.user import LoginSchema, UserSchema, login_schema, user_schema
 
 __all__ = [
-    LoginSchema,
-    RoleSchema,
-    UserSchema,
+    'LoginSchema',
+    'login_schema',
+    'RoleSchema',
+    'UserSchema',
+    'user_schema'
 ]
