+++ server/app/schemas/user.py
@@ -1,6 +1,6 @@
 """Serialized schemas for User model."""
+import uuid
 
-from flask_jwt_extended import create_access_token
 from marshmallow import Schema, fields, post_load
 from werkzeug.security import generate_password_hash
 
@@ -12,20 +12,27 @@ from app.utilities import generate_access_token
 class UserSchema(Schema):
     """Serialized user schema."""
 
-    id = fields.String(dump_only=True)
+    name = fields.String(required=True)
     email = fields.Email(required=True)
+    role = fields.Nested(RoleSchema, required=True)
+
+    # load only
     password = fields.String(load_only=True, required=True)
+
+    # dump only
+    id = fields.String(dump_only=True)
     date_registered = fields.Date(dump_only=True)
     auth_token = fields.String(dump_only=True)
-    role = fields.Nested(RoleSchema, required=True)
-    latitude = fields.Float()
-    longitude = fields.Float()
+    unique_identity = fields.String(dump_only=True)
 
     @post_load
     def create_user(self, data, **kwargs):
         """Create user after load."""
-        token = generate_access_token(identity=data['email'])
+        unique_identity = str(uuid.uuid4())
+        token = generate_access_token(identity=unique_identity)
         user_data = {
+            'unique_identity': unique_identity,
+            'name': data['name'],
             'email': data['email'],
             'role': data['role'],
             'hashed_password': generate_password_hash(data['password']),
@@ -34,8 +41,14 @@ class UserSchema(Schema):
         return User(**user_data)
 
 
+user_schema = UserSchema()
+
+
 class LoginSchema(Schema):
     """Serialized login schema."""
 
     email = fields.Email(load_only=True, required=True)
     password = fields.String(load_only=True, required=True)
+
+
+login_schema = LoginSchema()
